Move _Config to tune.py
